Move color_background_instead_of_text checkbox to Sorting tab
- Moved 'Apply mod coloring to background instead of text' checkbox
  from the Advanced settings tab to the Sorting settings tab
- Removed color_background logic from advanced_tab_controller.py
  (signals, setChecked/isChecked, helper methods, EventBus usage)
- Added color_background logic to sorting_tab_controller.py
  (signals, setChecked/isChecked, coloring mode change handlers)
- Reordered color_background_instead_of_text_toggle in settings.py
  to live alongside render_unity_rich_text in the Sorting section
- Updated docstrings to reflect the new responsibility split
