+++ app/views/settings_dialog.py
@@ -817,6 +817,14 @@ This basically preserves your mod coloring, user notes etc. for this many second
         )
         modlist_option_group_box_layout.addWidget(self.render_unity_rich_text_checkbox)
 
+        # Mod coloring checkbox
+        self.color_background_instead_of_text_checkbox = QCheckBox(
+            self.tr("Apply mod coloring to background instead of text")
+        )
+        modlist_option_group_box_layout.addWidget(
+            self.color_background_instead_of_text_checkbox
+        )
+
         # Download missing mods checkbox
         self.download_missing_mods_checkbox = QCheckBox(
             self.tr("Download missing mods automatically")
@@ -1506,11 +1514,6 @@ This basically preserves your mod coloring, user notes etc. for this many second
             self.tr("Check for mod updates on refresh")
         )
         group_layout.addWidget(self.show_mod_updates_checkbox)
-
-        self.color_background_instead_of_text_checkbox = QCheckBox(
-            self.tr("Apply mod coloring to background instead of text")
-        )
-        group_layout.addWidget(self.color_background_instead_of_text_checkbox)
 
         self.update_databases_on_startup_checkbox = QCheckBox(
             self.tr("Update databases on startup")
